small tweaks + problem with tooltip
